+++ resources/views/pages/partials/blocks/column_item.blade.php
@@ -15,7 +15,7 @@
         $statValue = $title;
     }
     $iconPresenter = app(\WebBlocks\Cms\Support\PublicRendering\PublicIconPresenter::class);
-    $iconClass = $iconPresenter->iconClass($block->publicContentIconSlug(), $block->publicIconTone());
+    $iconClass = $iconPresenter->iconClass($block->publicContentIconSlug(), $block->publicIconTone(), $block->publicIconSize());
     $badgeLabel = $block->publicBadgeLabel();
     $badgeClass = $iconPresenter->badgeClass($block->publicBadgeTone());
     $itemUrl = $block->localizedPublicUrl($block->url);