+++ resources/views/pages/partials/blocks/content_header.blade.php
@@ -6,7 +6,7 @@
     $headerClass = collect(['wb-content-header', $alignmentClass, $block->publicBackgroundMediaClass()])->filter()->implode(' ');
     $backgroundStyle = $block->publicBackgroundMediaStyle();
     $iconPresenter = app(\WebBlocks\Cms\Support\PublicRendering\PublicIconPresenter::class);
-    $iconClass = $iconPresenter->iconClass($block->publicContentIconSlug(), $block->publicIconTone());
+    $iconClass = $iconPresenter->iconClass($block->publicContentIconSlug(), $block->publicIconTone(), $block->publicIconSize());
     $badgeLabel = $block->publicBadgeLabel();
     $badgeClass = $iconPresenter->badgeClass($block->publicBadgeTone());